Delete Unecessary & Restore Necessary Files BC of Renameing of Dashboard
