+++ app/views/dashboards/show.html.erb
@@ -1,0 +1,38 @@
+<div class="container">
+  <div class="row">
+    <div class="col-1"></div>
+      <div class="col-10">
+        <div class="row mt-5">
+          <div class="col-12 d-flex justify-content-center">
+            <%= image_tag(@user.avatar, class:"avatar") if @user.avatar.attached?  %>
+          </div>
+          <div class="col-12 d-flex justify-content-center mt-3">
+            <h1><%= @user.full_name %></h1>
+          </div>
+          <div class="col-12 d-flex justify-content-center mb-3">
+            <h2><%= @user.body %></h2>
+        </div>
+      </div>
+    <div class="col-1"></div>
+      <div class="row">
+        <div class="col-1"></div>
+          <div class="col-10">
+          <% @links.each do |link|  %>
+            <% unless link.url.empty? %>
+              <div class="row">
+                <div class="col-12 d-flex justify-content-center">
+                  <%= link_to track_link_click_path(link), target:"_blank", class:"linkable btn btn-secondary" do %>
+                    <span class="linkable-span linkable-font"><%= link.title %></span>
+                  <% end %>
+                </div>
+              </div>
+            <% else %>
+            <%= nil %>
+            <% end %>
+          <% end %>
+          </div>
+        <div class="col-1"></div>
+      </div>
+    </div>
+  </div>
+</div>
