--- app/views/pages/contact.html.erb
@@ -1,2 +1,0 @@
-<h1>Pages#contact</h1>
-<p>Find me in app/views/pages/contact.html.erb</p>
