--- app/views/pages/home.html.erb
@@ -1,2 +1,0 @@
-<h1>Pages#home</h1>
-<p>Find me in app/views/pages/home.html.erb</p>
