--- test/controllers/dashboard_controller_test.rb
@@ -1,8 +1,0 @@
-require "test_helper"
-
-class DashboardControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
-    get dashboard_index_url
-    assert_response :success
-  end
-end
