+++ test/controllers/dashboards_controller_test.rb
@@ -1,0 +1,8 @@
+require "test_helper"
+
+class DashboardsControllerTest < ActionDispatch::IntegrationTest
+  test "should get index" do
+    get dashboard_index_url
+    assert_response :success
+  end
+end
